added enemy debuffs, modified the debuff logic.
Changed the value types of the enemy speed and HP from int to double

+++ TowerDefenseGame/TowerDefenseGame/Geometry/Point.cs
@@ -40,7 +40,7 @@
             } 
         }
 
-        public static void HandleMovement(Point objectPosition, Point targetPosition, int speed)
+        public static void HandleMovement(Point objectPosition, Point targetPosition, double speed)
         {
             //Point temp = new Point(objectPosition.X, objectPosition.Y + 5);
             //
